% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mandelbrot}
\alias{mandelbrot}
\title{Calculate and display Mandelbrot fractal}
\usage{
mandelbrot(width = 78L, height = 44L)
}
\arguments{
\item{width}{resolution}

\item{height}{resolution}
}
\description{
See the title; uses ordered dynamic scheduling for parallelism.
}
