% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{table_multiple_threads}
\alias{table_multiple_threads}
\title{Initialize an array of doubles in parallel}
\usage{
table_multiple_threads(size)
}
\arguments{
\item{size}{size of array}
}
\description{
Initialize an array (sin function) in parallel.
}
