% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{table_serial}
\alias{table_serial}
\title{Initialize an array of doubles in serial}
\usage{
table_serial(size)
}
\arguments{
\item{size}{size of array}
}
\description{
Initialize an array (sin function) in serial. This is for comparison purposes.
}
