% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{table_simd}
\alias{table_simd}
\title{Initialize an array of doubles with SIMD (single thread; vectorized)}
\usage{
table_simd(size)
}
\arguments{
\item{size}{size of array}
}
\description{
Initialize an array (sin function) with SIMD. Please note this probably won't
be any faster than serial without a parallel floating point library (AMD and Intel have some).
}
